Se elimina validacion de token y se mueven metodos a clases de utileria
