+++ src/main/java/mx/mexicocovid19/plataforma/controller/AyudaRestController.java
@@ -25,7 +25,6 @@ import mx.mexicocovid19.plataforma.controller.mapper.AyudaMapper;
 import mx.mexicocovid19.plataforma.exception.PMCException;
 import mx.mexicocovid19.plataforma.model.entity.Ayuda;
 import mx.mexicocovid19.plataforma.service.AyudaService;
-import mx.mexicocovid19.plataforma.util.ErrorEnum;
 
 /**
  * Created by betuzo on 14/05/15.
@@ -63,10 +62,7 @@ public class AyudaRestController {
         String username = jwtTokenUtil.getUsernameFromToken(token);
         
         ResponseEntity<AyudaDTO> response = new ResponseEntity<AyudaDTO>(HttpStatus.BAD_REQUEST);
-		if ( username == null ) {
-			throw new PMCException(ErrorEnum.ERR_AUTH_TOKEN_REQUERIDO, AyudaRestController.class.getName());
-		}
-		
+        
 		Ayuda createAyuda = ayudaService.createAyuda(AyudaMapper.from(ayudaDTO), username, request.getContextPath());
 		response = new ResponseEntity<AyudaDTO>(AyudaMapper.from(createAyuda), HttpStatus.OK);
