+++ src/main/java/mx/mexicocovid19/plataforma/util/Utils.java
@@ -1,10 +1,6 @@
 package mx.mexicocovid19.plataforma.util;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
-import java.util.TimeZone;
 
 import org.owasp.encoder.Encode;
 import org.springframework.http.HttpStatus;
@@ -24,12 +20,6 @@ public final class Utils {
 	/** constant string ERROR_TAG. */
 	private static final String ERROR_TAG = "ERROR FOUND || ";
 
-	/** {@link String} DATE_FORMAT formato para la salida de SimpleDateFormat. */
-	private static final String DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss:SSS'Z'";
-
-	/** {@link String} TIMEZONE_MEXICO Zona horaria de mexico centro. */
-	public static final String TIMEZONE_MEXICO = "America/Mexico_City";
-
 	/**
 	 * Nueva instancia utils.
 	 */
@@ -37,43 +27,6 @@ public final class Utils {
 		throw new UnsupportedOperationException("No instanciar la clase");
 	}
 
-	/**
-	 * This function allows get date in format ISO 8601.
-	 *
-	 * @return {@link String} - Date in format ISO 8601
-	 */
-	public static String getDateIso() {
-		/**
-		 * se crea la fecha tomada desde el sistema
-		 */
-		Date date = new Date(System.currentTimeMillis());
-		SimpleDateFormat sdf;
-		sdf = new SimpleDateFormat(DATE_FORMAT);
-		/**
-		 * se configura el timezone a la ciudad de Mexico
-		 */
-		sdf.setTimeZone(TimeZone.getTimeZone(TIMEZONE_MEXICO));
-		return sdf.format(date);
-	}
-
-	/**
-	 * Obtener el objeto: date isos.
-	 *
-	 * @return El objeto: date isos
-	 */
-	public static Date getDateIsos() {
-		Date date;
-		try {
-			/**
-			 * se crea una fecha con el ISO
-			 */
-			date = new SimpleDateFormat(DATE_FORMAT).parse(getDateIso());
-		} catch (ParseException e) {
-			Utils.printError(e, Utils.class.getName(), e.getMessage());
-			date = new Date();
-		}
-		return date;
-	}
 
 	/**
 	 * Prints the error binding.
